+++ cmd/completion/context.go
@@ -1,0 +1,18 @@
+package completion
+
+import (
+	"github.com/metal-stack/metalctl/pkg/api"
+	"github.com/spf13/cobra"
+)
+
+func (c *Completion) ContextListCompletion(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+	ctxs, err := api.GetContexts()
+	if err != nil {
+		return nil, cobra.ShellCompDirectiveError
+	}
+	var names []string
+	for name := range ctxs.Contexts {
+		names = append(names, name)
+	}
+	return names, cobra.ShellCompDirectiveNoFileComp
+}
